Sound
At time of submission, I have panel and all the client files in a separate folder with the sound file just in the java project folder, and everything else in a different folder with the sound file in the src folder. It should work in one folder except the sound will probably have to be in both the src and the project folder because I call it differently in the client. It works running both on one computer, but I've only tested it once on two computers and it didn't work. I put in a potential fix so it might work, but I haven't tested it yet.